Debugging: Class: Attica ---> Serious changes in Constructor | Class: Segment ---> Mainly on pass() method | Class: Entrance ---> Seriously important fix in insert_from_tolls() method

+++ Final/List.cpp
@@ -22,12 +22,17 @@ bool List :: is_empty() const{
 }
 
 void List :: enter(Vehicle* vehicle){
-	count++;
 	ListNode* new_node = new ListNode(vehicle,tail);
-	//cout << "Print from List Enter" << endl;
+	
+	if(is_empty()){
+		head->next = new_node;
+	}
+
+	count++;
+	
+
 	tail->next->next = new_node;	//the previously last node will point to the new last node
 	tail->next = new_node;	//the tail-node wiil also point to the new last node
-	//cout << "2.Print from List Enter" << endl;
 }
 
 void List :: exit(){
@@ -46,7 +51,7 @@ Vehicle* List :: pass(const bool& destroy){
 	if(!is_empty()){
 		//ListNode* ptr = head;	//initialize with head
 		for(ListNode* ptr = head; ptr->next != tail;ptr = ptr->next){
-			while(ptr->next->vehicle->is_ready() || (destroy == true)){ //Using a while loop instead of an if statement,because otherwise it would never delete consecutive nodes
+			if(ptr->next->vehicle->is_ready() || (destroy == true)){ //Using an if statement instead of a while loop,because one value gets returned
 				return delete_next(ptr->next);
 			}
 		}
@@ -56,22 +61,35 @@ Vehicle* List :: pass(const bool& destroy){
 
 int List :: get_ready_ones() const {
 	int ready_count = 0;
+	//cout << "There are " << get_count() << " vehicles" << endl;
 	if(!is_empty()){
+		//cout << "Ready count : " << ready_count << endl;
 		//ListNode* ptr = head;	//initialize with head
+		int i = 0;
 		for(ListNode* ptr = head; ptr->next != tail;ptr = ptr->next){
+		//	cout << i<< ".Vehicle: " << ptr->next->vehicle << endl;
 			if(ptr->next->vehicle->is_ready()){
+		//		cout << "Ready count : " << ready_count << endl;
 				ready_count++;
 			}
+			i++;
 		}
 	}
 	return ready_count;
 }
 
 void List :: set_ready(){
+	cout << "1.Print FROM THE LIST:: SET_READY()" << endl;
+	
 	int unready_ones = count - get_ready_ones();
+	
+	cout << "2.Print FROM THE LIST:: SET_READY()" << endl;
+	
 	int unready_ones_that_will_get_ready = unready_ones * percent;
 
 	ListNode* ptr = head->next;
+	
+	cout << "3.Print FROM THE LIST:: SET_READY()" << endl;
 	
 	for(int i = 0;i < unready_ones_that_will_get_ready;i++){
 		if(!ptr->vehicle->is_ready()){
@@ -91,11 +109,17 @@ Vehicle* List :: delete_next(ListNode* wanted_one){
 		//cout << "2nd print from delete next." << endl;
 	
 		count--;
+
+		if(is_empty()){
+			tail->next = head;
+			head->next = tail;
+		}
+
 		Vehicle* to_be_returned = ptr->next->vehicle;	//wanted node's vehicle
 		ListNode* second_node = ptr->next->next;	//the wanted node's next
 		delete ptr->next;	//delete the wanted node
 		ptr->next = second_node;	//connect the nodes
-		//to_be_returned->set_unready();	//set unready the vehicle that will be returned
+		to_be_returned->set_unready();	//set unready the vehicle that will be returned
 		return to_be_returned;
 		
 	}
